Clamp and offset floating colour options
Clamp positions within bounds of the window to prevent the mismatch
between the input and its selection mirror from having too terrible of
consequences. Also added an offset to the leading tail, derived from the
distance from the input's center. Applied the inverse offset to the
wrapper's position to keep the tail centred (and offset the menu
instead)