Open and close colour picker overlay conditionally
Colour options menu is now properly closed when colour picker button is
clicked or focus is removed from input. Clicking the colour picker
button now opens the colour picker menu, which closes when the user
cancels or confirms the colour picked. Also added the ability to close
the picker by clicking out of the menu.