Fix and refactor menu to properly open colour picker
Replaced a 'blur' event on the input with an invisible overlay rendered
with the options menu that cancels the selection when the menu isn't
interacted with. Previous commit broke the button by removing it from
the DOM on the input blurring before its own event could be handled.
Also re-ordered elements to ensure a proper stacking context for the
colour picker.