Apply picked colour to internal list of ranges
Colour emitted from the colour picker's `colour-set` event is now
properly parsed and added to the internal array for tracking the
colourised parts of the message. Overlapping ranges are also cropped
short, with precedence given to the new range. Added a couple helper
methods to the substring data structure to make the range parsing more
readable.