Generate message preview using render functions programmatically
Wanted to do this using a more "modern" VueJS way (with the auto script
setup rather than the export default), but this is the only way to use
the render functions. New component in the works to properly preview the
colourised changes being applied; it will replace the current input
mirror once complete.

Currently iterates through a sorted array to apply the spans, but
doesn't properly override the user's selection yet. The array passed to
it is also unsorted right now, so the colour application method will
need to be fixed (not the component's problem). For this component, the
contents of the input will need to be properly stored, so a new useState
has also been declared.