Add function to update colouredRanges on input changes
This new function adjusts the indices of the coloured ranges based on
new changes made to the input— both additions and removals, thanks to
the string difference util created a few commits ago. In essence, it is
supposed to extend and contract coloured sections when characters are
added or removed, respectively.