Add user selection visualization to MessagePreview
MessagePreview utilises the now-standalone function
adjustColoursAroundRange() to properly move the ranges of colours out of
the way. The space made is filled up with a span styled like a user
selection. In order to prevent the component from mutating the parent's
colour ranges, new clone() functions for IndexRange and ColouredRange
have been created to remove Vue's proxy reactivity.