Fix adjustColoursAroundRange() inserting ranges unsorted
When adjustColoursAroundRange() split a coloured range in two to wrap
around a nested range, it pushed the new second part of the colour to
the end of the array, causing the colours to appear out of order. To
fix, the second part of the colour is now inserted into the right index
using a splice instead of a push.