Integrate input width and scroll with MessagePreview
Officially added MessagePreview to the main Editor component. The
preview component now takes in width and scroll props, updating with the
main input's when passed through. To have more direct control over the
width of the input, the code to resize from input-resize.ts has been
moved directly into the editor's resize handler; the one that was
already calling it.