Optimise colourising algorithm for exported byte size
To minimise the amount of bytes used by the colourising algorithm,
ranges should be simplified wherever possible. For what needs to be
changed, this means removing those the same as the default text colour
and merging the same colours together.
Editor passes applyColour() the default chat colour, and the algorithm
will internally merge existing colours with the new range if the colours
match— if it's a ColouredRange (IndexRanges are treated the same, since
no colour lol).